requirements.txt
flask==3.1.0
line-bot-sdk==3.14.3
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.164.0
python-dotenv==1.1.0
